+++ systemtest/src/main/java/io/strimzi/systemtest/TestConstants.java
@@ -206,9 +206,7 @@ public interface TestConstants {
     /**
      * Feature gate related constants
      */
-    String DONT_USE_KRAFT_MODE = "-UseKRaft";
-    // kept for upgrade/downgrade tests in KRaft
-    String USE_KRAFT_MODE = "+UseKRaft";
+    // No Feature gates kept for STs at this moment
 
     /**
      * Default value which allows execution of tests with any tags